@@ -2,25 +2,25 @@
   "name": "ryoh",
   "version": "0.0.0",
   "scripts": {
-    "dev": "vite",
+    "dev": "vite --host 0.0.0.0",
     "build": "tsc && vite build",
     "preview": "vite preview"
   },
   "dependencies": {
-    "@mantine/core": "^4.2.5",
-    "@mantine/hooks": "^4.2.5",
+    "@mantine/core": "^4.2.6",
+    "@mantine/hooks": "^4.2.6",
+    "@tanstack/react-location": "^3.7.4",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
-    "react-icons": "^4.3.1",
-    "react-router-dom": "6.3.0",
-    "react-youtube": "^9.0.1",
-    "swr": "^1.3.0"
+    "react-query": "^3.39.0",
+    "react-youtube": "^9.0.2",
+    "tabler-icons-react": "^1.48.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.9",
-    "@types/react-dom": "^18.0.4",
+    "@types/react-dom": "^18.0.5",
     "@vitejs/plugin-react": "^1.3.2",
-    "typescript": "^4.6.4",
+    "typescript": "^4.7.2",
     "vite": "^2.9.9"
   }
 }
