@@ -1,17 +1,11 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import { BrowserRouter } from "react-router-dom";
-import { MantineProvider } from "@mantine/core";
-import "./index.css";
-import App from "./App";
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import { App } from "./components/App";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <BrowserRouter>
-      <MantineProvider theme={{ fontFamily: "'Noto Sans JP', sans-serif;" }}>
-        <App />
-      </MantineProvider>
-    </BrowserRouter>
-  </React.StrictMode>,
-  document.getElementById("root")
+const root = createRoot(document.getElementById("root")!);
+
+root.render(
+  <StrictMode>
+    <App />
+  </StrictMode>
 );
